Apply diffs between ArduinoCore-Samd 1.6.2 and 1.6.5.

+++ cores/arduino/Uart.cpp
@@ -34,7 +34,7 @@ void Uart::begin(unsigned long baudrate)
   begin(baudrate, (uint8_t)SERIAL_8N1);
 }
 
-void Uart::begin(unsigned long baudrate, uint8_t config)
+void Uart::begin(unsigned long baudrate, uint16_t config)
 {
 	uint8_t parity  =  config && HARDSER_PARITY_MASK;
 	uint8_t stopbit = (config && HARDSER_STOP_BIT_MASK) >> 8;
